requirements.txt
torchvision
pytorch-lightning
pandas
pillow
pytest
matplotlib
typeguard
torchtyping
deepposekit
imgaug
sklearn
hydra-core